reporting/article-retrieval/requirements.txt
streamlit 
matplotlib
google-cloud-secret-manager
numpy
pandas
duckdb==1.3.2
google-genai==1.25.0
pinecone
st-annotated-text